MAJOR REFACTOR 2: Validation, Error Handling, Endpoints, Docs
Backend:
- Extended input data validation for administrator and user endpoints
- Removed some of the try/catch blocks because of global exception handlers
- Added Telegram Bot Token validation by Telegram Bot API
- Removed formatted datetime and now using only datetime.now(timezone.utc) - forgot about this a long time ago

Frontend:
- Added new API responses from the Backend
- Changed timestamps formatting to locale string
- Added "Please login again!" responses inside frontend requests where some of them were missing

+++ asklvkaszus-backend/asklvkaszus/actions/admin/purge_question.py
@@ -1,30 +1,27 @@
 from ...extensions import csrf, sql
-from flask import current_app, request, jsonify
+from ...modules.fields import safe_get
+from flask import request, jsonify
 from ...models.questions import Questions
-import traceback
 
 def admin_purge_question(identity):
-    data = request.get_json()
-    question_id = data.get('question_id')
-
     csrf.protect()
 
-    try:
-        question = Questions.query.get(question_id)
+    data = request.get_json()
 
-        if not question:
-            return jsonify(error='Question with selected ID does not exist.'), 404
+    if not data:
+        return jsonify(error="Invalid JSON payload!"), 400
 
-        sql.session.delete(question)
-        sql.session.commit()
+    question_id = safe_get(data, 'question_id', str)
 
-        return jsonify(success='Question has been purged successfully!'), 200
+    if not question_id:
+        return jsonify(error='Please provide a Question ID!'), 400
 
-    except Exception as e:
-        current_app.logger.error("An error occured inside asklvkaszus/actions/admin/purge_question module: {e}")
-        traceback.print_exc()
+    question = Questions.query.get(question_id)
 
-        return jsonify(error='An error occurred while purging specified question! Try again later.'), 500
+    if not question:
+        return jsonify(error='Question with selected ID does not exist.'), 404
 
-    finally:
-        sql.session.close()
+    sql.session.delete(question)
+    sql.session.commit()
+
+    return jsonify(success='Question has been purged successfully!'), 200
